# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InfrastructureProjectProps": {
        "properties": {
          "budget_sek": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Sek"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "impact_radius_m": {
            "default": 1000.0,
            "title": "Impact Radius M",
            "type": "number"
          },
          "metadata_json": {
            "additionalProperties": true,
            "title": "Metadata Json",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectType"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "default": "manual",
            "title": "Source",
            "type": "string"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "InfrastructureProjectProps",
        "type": "object"
      },
      "NearbyProject": {
        "description": "Ett infrastrukturprojekt i närheten av en fastighet.",
        "properties": {
          "distance_m": {
            "description": "Avstånd i meter från fastigheten till projektet",
            "title": "Distance M",
            "type": "number"
          },
          "project": {
            "$ref": "#/components/schemas/InfrastructureProjectProps"
          },
          "within_impact_radius": {
            "description": "Om fastigheten ligger inom projektets påverkansradie",
            "title": "Within Impact Radius",
            "type": "boolean"
          }
        },
        "required": [
          "project",
          "distance_m",
          "within_impact_radius"
        ],
        "title": "NearbyProject",
        "type": "object"
      },
      "NearbyProjectsResponse": {
        "properties": {
          "max_distance_m": {
            "title": "Max Distance M",
            "type": "number"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/NearbyProject"
            },
            "title": "Projects",
            "type": "array"
          },
          "property_id": {
            "title": "Property Id",
            "type": "integer"
          }
        },
        "required": [
          "property_id",
          "max_distance_m"
        ],
        "title": "NearbyProjectsResponse",
        "type": "object"
      },
      "ProjectStatus": {
        "enum": [
          "planerad",
          "pågående",
          "avslutad"
        ],
        "title": "ProjectStatus",
        "type": "string"
      },
      "ProjectType": {
        "enum": [
          "väg",
          "järnväg",
          "kollektivtrafik",
          "bro",
          "tunnel",
          "cykelväg",
          "övrigt"
        ],
        "title": "ProjectType",
        "type": "string"
      },
      "PropertyCollection": {
        "properties": {
          "features": {
            "items": {
              "$ref": "#/components/schemas/PropertyFeature"
            },
            "title": "Features",
            "type": "array"
          },
          "numberMatched": {
            "description": "Totalt antal träffar för frågan",
            "title": "Numbermatched",
            "type": "integer"
          },
          "numberReturned": {
            "description": "Antal features i detta svar",
            "title": "Numberreturned",
            "type": "integer"
          },
          "type": {
            "const": "FeatureCollection",
            "default": "FeatureCollection",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "numberMatched",
          "numberReturned"
        ],
        "title": "PropertyCollection",
        "type": "object"
      },
      "PropertyCreate": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "area_sqm": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Sqm"
          },
          "assessed_value_sek": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessed Value Sek"
          },
          "building_year": {
            "anyOf": [
              {
                "maximum": 2200.0,
                "minimum": 1000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Building Year"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "county": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "County"
          },
          "designation": {
            "minLength": 1,
            "title": "Designation",
            "type": "string"
          },
          "geometry": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "GeoJSON-geometri (Polygon eller MultiPolygon) i WGS84",
            "title": "Geometry"
          },
          "living_area_sqm": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Living Area Sqm"
          },
          "metadata_json": {
            "additionalProperties": true,
            "title": "Metadata Json",
            "type": "object"
          },
          "municipality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Municipality"
          },
          "owner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Name"
          },
          "owner_org_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Org Number"
          },
          "postal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Code"
          },
          "property_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PropertyType"
              },
              {
                "type": "null"
              }
            ]
          },
          "zoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zoning"
          }
        },
        "required": [
          "designation"
        ],
        "title": "PropertyCreate",
        "type": "object"
      },
      "PropertyFeature": {
        "properties": {
          "geometry": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry"
          },
          "properties": {
            "$ref": "#/components/schemas/PropertyProps"
          },
          "type": {
            "const": "Feature",
            "default": "Feature",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "properties"
        ],
        "title": "PropertyFeature",
        "type": "object"
      },
      "PropertyProps": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "area_sqm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Sqm"
          },
          "assessed_value_sek": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessed Value Sek"
          },
          "building_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Building Year"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "county": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "County"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "designation": {
            "title": "Designation",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "living_area_sqm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Living Area Sqm"
          },
          "metadata_json": {
            "additionalProperties": true,
            "title": "Metadata Json",
            "type": "object"
          },
          "municipality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Municipality"
          },
          "owner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Name"
          },
          "owner_org_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Org Number"
          },
          "postal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Code"
          },
          "property_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PropertyType"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "zoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zoning"
          }
        },
        "required": [
          "id",
          "designation"
        ],
        "title": "PropertyProps",
        "type": "object"
      },
      "PropertyType": {
        "enum": [
          "bostad",
          "kontor",
          "handel",
          "industri",
          "utbildning",
          "villa"
        ],
        "title": "PropertyType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API för visualisering av svenska fastigheter och infrastrukturprojekt. Geodata lagras i PostGIS och serveras som GeoJSON.",
    "title": "Fastighetsvisualiserare API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/properties": {
      "get": {
        "description": "Fastigheter som GeoJSON FeatureCollection, med filter och paginering.",
        "operationId": "list_properties_api_v1_properties_get",
        "parameters": [
          {
            "description": "Filtrera på kommun (kan upprepas)",
            "in": "query",
            "name": "municipality",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrera på kommun (kan upprepas)",
              "title": "Municipality"
            }
          },
          {
            "description": "Filtrera på fastighetstyp (kan upprepas)",
            "in": "query",
            "name": "property_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/PropertyType"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrera på fastighetstyp (kan upprepas)",
              "title": "Property Type"
            }
          },
          {
            "description": "Lägsta taxeringsvärde (SEK)",
            "in": "query",
            "name": "min_value",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Lägsta taxeringsvärde (SEK)",
              "title": "Min Value"
            }
          },
          {
            "description": "Högsta taxeringsvärde (SEK)",
            "in": "query",
            "name": "max_value",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Högsta taxeringsvärde (SEK)",
              "title": "Max Value"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 2000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Avgränsningsruta i WGS84: väst,syd,öst,norr",
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Avgränsningsruta i WGS84: väst,syd,öst,norr",
              "title": "Bbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyCollection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Properties",
        "tags": [
          "properties"
        ]
      },
      "post": {
        "operationId": "create_property_api_v1_properties_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyFeature"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Property",
        "tags": [
          "properties"
        ]
      }
    },
    "/api/v1/properties/search": {
      "get": {
        "description": "Fritextsök på beteckning, adress, ägare, stad och kommun.",
        "operationId": "search_properties_api_v1_properties_search_get",
        "parameters": [
          {
            "description": "Sökterm",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Sökterm",
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyCollection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Properties",
        "tags": [
          "properties"
        ]
      }
    },
    "/api/v1/properties/{property_id}": {
      "get": {
        "operationId": "get_property_api_v1_properties__property_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "title": "Property Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyFeature"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Property",
        "tags": [
          "properties"
        ]
      }
    },
    "/api/v1/properties/{property_id}/nearby-projects": {
      "get": {
        "description": "Infrastrukturprojekt nära en fastighet, närmast först (ST_DWithin).",
        "operationId": "nearby_projects_api_v1_properties__property_id__nearby_projects_get",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "title": "Property Id",
              "type": "integer"
            }
          },
          {
            "description": "Maximalt avstånd i meter",
            "in": "query",
            "name": "max_distance_m",
            "required": false,
            "schema": {
              "default": 5000,
              "description": "Maximalt avstånd i meter",
              "exclusiveMinimum": 0,
              "maximum": 50000,
              "title": "Max Distance M",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NearbyProjectsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Nearby Projects",
        "tags": [
          "properties"
        ]
      }
    }
  }
}